Fix flake8 issues in eudemo.ivc.panelling
